+++ src/app/bases/VersionQueryBuilder.ts
@@ -7,16 +7,18 @@ import { IQueryBuilder } from './IQueryBuilder';
 export class VersionQueryBuilder<TEntity, TModel, TPk extends PkType, TUk = NameUk> 
 	implements IQueryBuilder<TEntity, TModel, TPk, TUk> {
 
+	private _pkProps: string[];
+
 	constructor(private _EntityClass: Newable) {
+		this._pkProps = this._EntityClass['idProp'];
 	}
-
-
+	
 	public buildCountAll(prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>, opts: it.RepositoryCountAllOptions): QueryBuilder<TEntity> {
 		return prevQuery.where('is_main', true);
 	}
 
 	public buildDeleteHard(pk: TPk, prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>): QueryBuilderSingle<number> {
-		return rawQuery.deleteById(this._toArr(pk, this._EntityClass['idProp']));
+		return rawQuery.deleteById(this._toArr(pk, this._pkProps));
 	}
 
 	public buildExists(props: TUk, prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>, opts: it.RepositoryExistsOptions): QueryBuilder<TEntity> {
@@ -24,7 +26,7 @@ export class VersionQueryBuilder<TEntity, TModel, TPk extends PkType, TUk = Name
 	}
 
 	public buildFind(pk: TPk, prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>, opts: it.RepositoryFindOptions = {}): QueryBuilder<TEntity> {
-		let q = rawQuery.findById(this._toArr(pk, this._EntityClass['idProp']));
+		let q = rawQuery.findById(this._toArr(pk, this._pkProps));
 		if (opts.version) {
 			q = q.where('version', opts.version);
 		} else {
@@ -38,7 +40,7 @@ export class VersionQueryBuilder<TEntity, TModel, TPk extends PkType, TUk = Name
 	}
 
 	public buildPatch(entity: TEntity, prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>, opts: it.RepositoryPatchOptions): QueryBuilder<number> {
-		return <any>rawQuery.patch(entity).whereComposite(this._EntityClass['idColumn'], '=', this._toArr(entity, this._EntityClass['idProp'])).where('is_main', true);
+		return <any>rawQuery.patch(entity).whereComposite(this._EntityClass['idColumn'], '=', this._toArr(entity, this._pkProps)).where('is_main', true);
 	}
 
 	public buildRecoverOpts(pk: TPk, prevOpts: it.RepositoryRecoverOptions, rawOpts: it.RepositoryRecoverOptions): it.RepositoryExistsOptions {
@@ -46,7 +48,7 @@ export class VersionQueryBuilder<TEntity, TModel, TPk extends PkType, TUk = Name
 	}
 
 	public buildUpdate(entity: TEntity, prevQuery: QueryBuilder<TEntity>, rawQuery: QueryBuilder<TEntity>, opts: it.RepositoryPatchOptions): QueryBuilder<number> {
-		return <any>rawQuery.update(entity).whereComposite(this._EntityClass['idColumn'], '=', this._toArr(entity, this._EntityClass['idProp'])).where('is_main', true);
+		return <any>rawQuery.update(entity).whereComposite(this._EntityClass['idColumn'], '=', this._toArr(entity, this._pkProps)).where('is_main', true);
 	}